SOLR-15337: cleanup CHANGES.txt
